implement partial chowimg extraction
When using Baba Is You's Assets.dat for testing, I noticed that
there are a few images that fail to extract and I do not quite know
what's going on in these entries after checking them in a hex editor.

It appears that I won't get around reverse engineering the binary if I
want to understand this format fully.